# src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Distributed Inventory API (Marketplace Brazil)
  version: 0.1.1
  description: |
    OpenAPI contract for a distributed inventory system with strong-consistency writes,
    reservation with TTL, CQRS (write canonical + read projections/cache), and idempotent
    mutations. Suitable for a marketplace with multiple stores/sellers across Brazil.
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging.api.example.com
    description: Staging
security:
  - bearerAuth: []

tags:
  - name: Items (Write)
  - name: Reservations (Write)
  - name: Inventory (Read)

paths:
  /items:
    post:
      tags: [Items (Write)]
      summary: Create an item (SKU)
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
            examples:
              default:
                value:
                  sku: "SKU-123"
                  name: "Headphone XYZ"
                  attributes:
                    color: "black"
                    brand: "ACME"
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: Resource location
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Inventory (Read)]
      summary: Search items (optionally scoped by store)
      operationId: searchItems
      parameters:
        - in: query
          name: storeId
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchItemsResponse' }

  /items/{sku}/adjust:
    put:
      tags: [Items (Write)]
      summary: Adjust inventory for an SKU in a store (conditional write)
      operationId: adjustInventory
      parameters:
        - $ref: '#/components/parameters/skuPath'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdjustRequest' }
      responses:
        '200':
          description: Inventory updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryRecord' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  /items/{sku}/reserve:
    post:
      tags: [Reservations (Write)]
      summary: Create a reservation for an SKU (with TTL)
      operationId: createReservation
      parameters:
        - $ref: '#/components/parameters/skuPath'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReserveRequest' }
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reservation' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  /items/{sku}/commit:
    post:
      tags: [Reservations (Write)]
      summary: Commit a reservation
      operationId: commitReservation
      parameters:
        - $ref: '#/components/parameters/skuPath'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommitRequest' }
      responses:
        '200':
          description: Committed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryRecord' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  /items/{sku}/release:
    post:
      tags: [Reservations (Write)]
      summary: Release a reservation
      operationId: releaseReservation
      parameters:
        - $ref: '#/components/parameters/skuPath'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReleaseRequest' }
      responses:
        '200':
          description: Released
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryRecord' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  /items/{sku}:
    get:
      tags: [Inventory (Read)]
      summary: Get item details + availability
      operationId: getItem
      parameters:
        - $ref: '#/components/parameters/skuPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GetItemResponse' }
        '404': { $ref: '#/components/responses/NotFound' }

  /availability:
    get:
      tags: [Inventory (Read)]
      summary: Batch read availability by SKUs
      operationId: getAvailability
      parameters:
        - in: query
          name: skus
          required: true
          schema:
            type: array
            items: { type: string }
            style: form
            explode: false
        - in: query
          name: storeId
          required: false
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GetAvailabilityResponse' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    skuPath:
      in: path
      name: sku
      required: true
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string }
    CorrelationId:
      in: header
      name: X-Correlation-Id
      required: false
      schema: { type: string }
  responses:
    BadRequest:
      description: Bad Request
    NotFound:
      description: Not Found
    Conflict:
      description: Conflict
  schemas:
    Item:
      type: object
      properties:
        sku: { type: string }
        name: { type: string }
        attributes:
          type: object
          additionalProperties: { type: string }
      required: [sku, name]
    InventoryRecord:
      type: object
      properties:
        sku: { type: string }
        storeId: { type: string }
        quantityAvailable: { type: integer }
        reserved: { type: integer }
        version: { type: integer }
        updatedAt: { type: string, format: date-time }
      required: [sku, storeId, quantityAvailable, reserved, version, updatedAt]
    Reservation:
      type: object
      properties:
        reservationId: { type: string }
        sku: { type: string }
        storeId: { type: string }
        qty: { type: integer }
        status:
          type: string
          enum: [CREATED, COMMITTED, RELEASED, EXPIRED]
        expiresAt: { type: string, format: date-time }
    CreateItemRequest:
      type: object
      properties:
        sku: { type: string }
        name: { type: string }
        attributes:
          type: object
          additionalProperties: { type: string }
    AdjustRequest:
      type: object
      properties:
        storeId: { type: string }
        delta: { type: integer }
        expectedVersion: { type: integer }
    ReserveRequest:
      type: object
      properties:
        qty: { type: integer }
        storeId: { type: string }
        ttlSeconds: { type: integer }
    CommitRequest:
      type: object
      properties:
        reservationId: { type: string }
    ReleaseRequest:
      type: object
      properties:
        reservationId: { type: string }
    GetItemResponse:
      type: object
      properties:
        item: { $ref: '#/components/schemas/Item' }
        availability:
          type: array
          items: { $ref: '#/components/schemas/InventoryRecord' }
    GetAvailabilityResponse:
      type: object
      properties:
        results:
          type: array
          items: { $ref: '#/components/schemas/InventoryRecord' }
    SearchItemsResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Item' }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }